小修正反映

+++ alu.h
@@ -5,8 +5,8 @@ unsigned int signExt(unsigned int argument);
 unsigned int sw(unsigned int rt, unsigned int address, unsigned int* memory);
 unsigned int lw(unsigned int address, unsigned int* memory);
 
-unsigned int fpuHide(unsigned int pc, unsigned int instruction, unsigned int* reg, unsigned int* fpreg, unsigned int* fpuNum, unsigned int* labelRec);
-unsigned int fpu(unsigned int pc, unsigned int instruction, unsigned int* reg, unsigned int* fpreg, int* flag, unsigned int* fpuNum, unsigned int* labelRec);
+unsigned int fpuHide(unsigned int pc, unsigned int instruction, unsigned int* reg, unsigned int* fpreg, unsigned long long* fpuNum, unsigned int* labelRec);
+unsigned int fpu(unsigned int pc, unsigned int instruction, unsigned int* reg, unsigned int* fpreg, int* flag, unsigned long long* fpuNum, unsigned int* labelRec);
 
 unsigned int sll(unsigned int rs, unsigned int shamt);
 unsigned int srl(unsigned int rs, unsigned int shamt);
